requirements.txt
keyboard
pyautogui
Pillow
pystray
google-genai